Remove __init__ and initialise Config correctly
